Extract uptime to its own package
- Centralize all uptime related logic into istanbul/uptime pkg.
- Use interface for storage to ease testing later
