+++ consensus/istanbul/utils.go
@@ -119,37 +119,6 @@ func GetEpochLastBlockNumber(epochNumber uint64, epochSize uint64) uint64 {
 	return epochNumber * epochSize
 }
 
-// GetUptimeMonitoringWindow retrieves the range [first block, last block] where uptime is to be monitored
-// for a give epoch. The range is inclusive.
-// First blocks of an epoch need to be skipped since we can't assess the last `lookbackWindow` block for validators
-// as those are froma different epoch.
-// Similarly, last block of epoch is skipped since we can't obtaine the signer for it; as they are in the next block
-func GetUptimeMonitoringWindow(epochNumber uint64, epochSize uint64, lookbackWindowSize uint64) (uint64, uint64) {
-	if epochNumber == 0 {
-		panic("no monitoring window for epoch 0")
-	}
-
-	epochFirstBlock, _ := GetEpochFirstBlockNumber(epochNumber, epochSize)
-	epochLastBlock := GetEpochLastBlockNumber(epochNumber, epochSize)
-
-	// first block to monitor:
-	// We need to wait for the completion of the first window with the start window's block being the
-	// 2nd block of the epoch, before we start tallying the validator score for epoch "epochNumber".
-	// We can't include the epoch's first block since it's aggregated parent seals
-	// is for the previous epoch's valset.
-	firstBlockToMonitor := epochFirstBlock + 1 + (lookbackWindowSize - 1)
-
-	// last block to monitor:
-	// We stop tallying for epoch "epochNumber" at the second to last block of that epoch.
-	// We can't include that epoch's last block as part of the tally because the epoch val score is calculated
-	// using a tally that is updated AFTER a block is finalized.
-	// Note that it's possible to count up to the last block of the epoch, but it's much harder to implement
-	// than couting up to the second to last one.
-	lastBlockToMonitor := epochLastBlock - 1
-
-	return firstBlockToMonitor, lastBlockToMonitor
-}
-
 func ValidatorSetDiff(oldValSet []ValidatorData, newValSet []ValidatorData) ([]ValidatorData, *big.Int) {
 	valSetMap := make(map[common.Address]bool)
 	oldValSetIndices := make(map[common.Address]int)
