+++ core/blockchain.go
@@ -21,7 +21,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"math"
 	"math/big"
 	mrand "math/rand"
 	"sort"
@@ -33,7 +32,8 @@ import (
 	"github.com/ethereum/go-ethereum/common/mclock"
 	"github.com/ethereum/go-ethereum/common/prque"
 	"github.com/ethereum/go-ethereum/consensus"
-	"github.com/ethereum/go-ethereum/consensus/istanbul"
+	"github.com/ethereum/go-ethereum/consensus/istanbul/uptime"
+	"github.com/ethereum/go-ethereum/consensus/istanbul/uptime/store"
 	"github.com/ethereum/go-ethereum/core/rawdb"
 	"github.com/ethereum/go-ethereum/core/state"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -1242,71 +1242,6 @@ func (bc *BlockChain) writeKnownBlock(block *types.Block) error {
 	return nil
 }
 
-// https://stackoverflow.com/questions/19105791/is-there-a-big-bitcount/32702348#32702348
-func bitCount(n *big.Int) int {
-	count := 0
-	for _, v := range n.Bits() {
-		count += popcount(uint64(v))
-	}
-	return count
-}
-
-// Straight and simple C to Go translation from https://en.wikipedia.org/wiki/Hamming_weight
-func popcount(x uint64) int {
-	const (
-		m1  = 0x5555555555555555 //binary: 0101...
-		m2  = 0x3333333333333333 //binary: 00110011..
-		m4  = 0x0f0f0f0f0f0f0f0f //binary:  4 zeros,  4 ones ...
-		h01 = 0x0101010101010101 //the sum of 256 to the power of 0,1,2,3...
-	)
-	x -= (x >> 1) & m1             //put count of each 2 bits into those 2 bits
-	x = (x & m2) + ((x >> 2) & m2) //put count of each 4 bits into those 4 bits
-	x = (x + (x >> 4)) & m4        //put count of each 8 bits into those 8 bits
-	return int((x * h01) >> 56)    //returns left 8 bits of x + (x<<8) + (x<<16) + (x<<24) + ...
-}
-
-// updateUptime receives the currently accumulated uptime for all validators in an epoch and proceeds to update it
-// based on which validators signed on the provided block by inspecting the block's parent bitmap
-//
-// It expects a blockNumber with the signatures bitmap from the previous block (parent seal)
-func updateUptime(uptime *istanbul.Uptime, blockNumber uint64, bitmap *big.Int, window uint64, epochNum uint64, epochSize uint64) *istanbul.Uptime {
-	if uptime == nil {
-		uptime = new(istanbul.Uptime)
-		// The number of validators is upper bounded by 3/2 of the number of 1s in the bitmap
-		// We multiply by 2 just to be extra cautious of off-by-one errors.
-		validatorsSizeUpperBound := uint64(math.Ceil(float64(bitCount(bitmap)) * 2))
-		uptime.Entries = make([]istanbul.UptimeEntry, validatorsSizeUpperBound)
-	}
-
-	firstBlockToMonitor, lastBlockToMonitor := istanbul.GetUptimeMonitoringWindow(epochNum, epochSize, window)
-
-	// signedBlockWindowLastBlockNum is just the previous block
-	signedBlockWindowLastBlockNum := blockNumber - 1
-	signedBlockWindowFirstBlockNum := signedBlockWindowLastBlockNum - (window - 1)
-
-	uptime.LatestBlock = blockNumber
-	for i := 0; i < len(uptime.Entries); i++ {
-		if bitmap.Bit(i) == 1 {
-			// update their latest signed block
-			uptime.Entries[i].LastSignedBlock = blockNumber - 1
-		}
-
-		// If we are within the validator uptime tally window, then update the validator's score
-		// if its last signed block is within the lookback window
-		if blockNumber >= firstBlockToMonitor && blockNumber <= lastBlockToMonitor {
-			lastSignedBlock := uptime.Entries[i].LastSignedBlock
-
-			// Note that the second condition in the if condition is not necessary.  But it does
-			// make the logic easier to understand.  (e.g. it's checking is lastSignedBlock is within
-			// the range [signedBlockWindowFirstBlockNum, signedBlockWindowLastBlockNum])
-			if lastSignedBlock >= signedBlockWindowFirstBlockNum && lastSignedBlock <= signedBlockWindowLastBlockNum {
-				uptime.Entries[i].ScoreTally++
-			}
-		}
-	}
-	return uptime
-}
-
 // WriteBlockWithState writes the block and all associated state to the database.
 func (bc *BlockChain) WriteBlockWithState(block *types.Block, receipts []*types.Receipt, logs []*types.Log, state *state.StateDB, emitHeadEvent bool) (status WriteStatus, err error) {
 	bc.chainmu.Lock()
@@ -1321,8 +1256,6 @@ func (bc *BlockChain) writeBlockWithState(block *types.Block, receipts []*types.
 	bc.wg.Add(1)
 	defer bc.wg.Done()
 
-	// We are going to update the uptime tally.
-	// TODO find a better way of checking if it's istanbul
 	randomCommitment := common.Hash{}
 	if istEngine, isIstanbul := bc.engine.(consensus.Istanbul); isIstanbul {
 
@@ -1330,33 +1263,11 @@ func (bc *BlockChain) writeBlockWithState(block *types.Block, receipts []*types.
 			log.Error("Found two blocks with same height", "old", hash, "new", block.Hash())
 		}
 
-		// The epoch's first block's aggregated parent signatures is for the previous epoch's valset.
-		// We can ignore updating the tally for that block.
-		if !istanbul.IsFirstBlockOfEpoch(block.NumberU64(), bc.chainConfig.Istanbul.Epoch) {
-			epochNum := istanbul.GetEpochNumber(block.NumberU64(), bc.chainConfig.Istanbul.Epoch)
-
-			// Get the bitmap from the previous block
-			extra, err := types.ExtractIstanbulExtra(block.Header())
-			if err != nil {
-				log.Error("Unable to extract istanbul extra", "func", "WriteBlockWithState", "blocknum", block.NumberU64(), "epoch", epochNum)
-				return NonStatTy, errors.New("could not extract block header extra")
-			}
-			signedValidatorsBitmap := extra.ParentAggregatedSeal.Bitmap
-
-			// Get the uptime scores
-			uptime := rawdb.ReadAccumulatedEpochUptime(bc.db, epochNum)
-
-			// We only update the uptime for blocks which are greater than the last block we saw.
-			// This ensures that we do not count the same block twice for any reason.
-			if uptime == nil || uptime.LatestBlock < block.NumberU64() {
-				// Update the uptime scores
-				uptime = updateUptime(uptime, block.NumberU64(), signedValidatorsBitmap, bc.chainConfig.Istanbul.LookbackWindow, epochNum, bc.chainConfig.Istanbul.Epoch)
-
-				// Write the new uptime scores
-				rawdb.WriteAccumulatedEpochUptime(bc.db, epochNum, uptime)
-			} else {
-				log.Trace("WritingBlockWithState with block number less than a block we previously wrote", "latestUptimeBlock", uptime.LatestBlock, "blockNumber", block.NumberU64())
-			}
+		// We are going to update the uptime tally.
+		uptimeMonitor := uptime.NewMonitor(store.New(bc.db), bc.chainConfig.Istanbul.Epoch, bc.chainConfig.Istanbul.LookbackWindow)
+		err = uptimeMonitor.ProcessBlock(block)
+		if err != nil {
+			return NonStatTy, err
 		}
 
 		blockAuthor, err := istEngine.Author(block.Header())
